added navbar based on my page design, made it a component based
