+++ scrybe_frontend/src/pages/UploadRecordings/UploadedNavbar/uploadedNavbar.module.scss
@@ -1,0 +1,176 @@
+.UploadedRecNavbar {
+  background: #ffffff;
+  height: 80px;
+  width: 100%;
+  display: flex;
+  justify-content: space-between;
+  padding: 1rem;
+  position: relative;
+
+  //toggle nav and logo
+  .UploadedRecNavbar_toggle {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    img {
+      margin-right: 1rem;
+    }
+    display: none;
+  }
+
+  // search input
+  .UploadedRecNavbar_inputwrap {
+    display: flex;
+    align-items: center;
+    width: 100%;
+    max-width: 400px;
+    height: 40px;
+    border: 1px solid #e0e0e0;
+    border-radius: 5px;
+    padding: 0 1rem;
+    -webkit-border-radius: 5px;
+    -moz-border-radius: 5px;
+    -ms-border-radius: 5px;
+    -o-border-radius: 5px;
+    .searchIcon {
+      color: #e0e0e0;
+      font-size: 1.2rem;
+      margin-right: 0.5rem;
+    }
+    input {
+      background: transparent;
+      border: none;
+      outline: none;
+      width: 100%;
+      height: 100%;
+      font-size: 1rem;
+      color: #a9a9a9;
+
+      ::placeholder {
+        font-family: "Kanit";
+        font-style: normal;
+        font-weight: 300;
+        font-size: 16px;
+        color: #a9a9a9;
+        line-height: 24px;
+        /* identical to box height, or 150% */
+
+        /* Grey/grey-200 */
+      }
+    }
+  }
+
+  // uset car and btn
+  .UploadedRecNavbar_user_btn {
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    .UploadedRecNavbar_user {
+      display: flex;
+      // align-items: center;
+      // justify-content: space-between;
+      margin-right: 1rem;
+      .UploadedRecNavbar_user_desktop {
+        display: flex;
+        // align-items: center;
+        // justify-content: space-between;
+        img {
+          margin-right: 1rem;
+        }
+        .UploadedRecNavbar_user_desktop_nameDetails {
+          width: 125px !important;
+          height: 48px;
+          // padding: 1rem 0;
+          position: relative;
+          .UploadedRecNavbar_user_desktop_name_arr {
+            display: flex;
+            .name {
+              font-family: "Kanit";
+              font-style: normal;
+              font-weight: 600;
+              font-size: 20px;
+              line-height: 20px;
+              margin-right: 0.5rem;
+              /* identical to box height, or 120% */
+
+              /* Grey/grey-500 */
+
+              color: #454545;
+            }
+          }
+          .workspace_name {
+            font-family: "Kanit";
+            font-style: normal;
+            font-weight: 300;
+            font-size: 14px;
+            line-height: 20px;
+            /* identical to box height, or 143% */
+
+            /* Grey/grey-200 */
+
+            color: #a9a9a9;
+          }
+        }
+      }
+    }
+
+    // user img mobile
+    .UploadedRecNavbar_user_mobile {
+      display: none;
+    }
+    .UploadedNavbarRec_btnwrap {
+      background: #006cff;
+      border-radius: 4px;
+      display: block;
+      padding: 12px 32px;
+      width: 143px;
+      display: flex;
+      height: 48px;
+      img {
+        margin-right: 0.5rem;
+        width: 14px;
+      }
+      .UploadedNavbarRec_btn {
+        display: block;
+        outline: none;
+        border: none;
+        background: none;
+        color: #ffffff;
+        font-family: "Kanit";
+        font-weight: 500;
+        font-size: 14px;
+      }
+    }
+  }
+}
+
+// responsive
+@media screen and (max-width: 768px) {
+  .UploadedRecNavbar {
+    // background: #000;
+    padding: 0.5rem;
+
+    .UploadedRecNavbar_toggle {
+      display: flex;
+    }
+    .UploadedRecNavbar_inputwrap {
+      display: none !important;
+    }
+    .UploadedRecNavbar_user_desktop {
+      display: none !important;
+    }
+    .UploadedNavbarRec_btnwrap {
+      display: none !important;
+    }
+    .UploadedRecNavbar_user_mobile {
+      display: block !important;
+      float: right;
+      position: relative;
+      img {
+        // position: absolute;
+        top: 1rem;
+        right: 1rem;
+      }
+    }
+  }
+}
